requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.24.0
starlette==0.27.0
pydantic==2.5.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
email-validator==2.1.0.post1
httpx==0.25.2
python-multipart==0.0.6
Pillow==10.1.0
python-docx==1.1.0
reportlab==4.0.8
arabic-reshaper==3.0.0
python-bidi==0.4.2
